requirements.txt
Flask==1.0.2
Flask-Uploads==0.2.1
Flask-WTF==0.14.2
future==0.17.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
Pillow==5.4.1
requests==2.21.0
urllib3==1.24.1
Werkzeug==0.14.1
WTForms==2.2.1
